requirements.txt
plotly
streamlit
pandas
Pillow  # for image handling 
google-cloud-aiplatform
vertexai
python-dotenv
google-cloud-discoveryengine
pypdf
google-genai